finish lab 4.2

+++ include/sst/sst.h
@@ -72,7 +72,7 @@ public:
   std::shared_ptr<Block> read_block(size_t block_idx);
 
   // 找到key所在的block的idx
-  size_t find_block_idx(const std::string &key);
+  int find_block_idx(const std::string &key);
 
   // 根据key返回迭代器
   SstIterator get(const std::string &key, uint64_t tranc_id);
